Multi-Walker/HARL_based/pyproject.toml
[project]
name = "harl"
version = "1.0.0"
description = "PyTorch implementation of HARL Algorithms"
readme = "README.md"
requires-python = ">=3.8,<3.9"
dependencies = [
  "torch==1.9",
  "pyyaml>=5.3.1",
  "tensorboard>=2.2.1",
  "tensorboardX",
  "setproctitle",
  "pettingzoo[mpe,sisl]>=1.24.3",
  "wandb>=0.18.7",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "rich>=13.9.4",
  "hydra-submitit-launcher>=1.2.0",
  "hydra-colorlog>=1.2.0",
  "supersuit>=3.9.3",
  "matplotlib>=3.7.5",
  "imageio>=2.35.1",
  "moviepy>=1.0.3",
  "python-dotenv>=1.0.1",
]

[project.urls]
homepage = "https://github.com/PKU-MARL/HARL"

[[tool.uv.index]]
name = "cstcloud"
url = "https://pypi.cstcloud.cn/simple"
default = true
